[VM] Break out values, loaded values, and natives into their own crate
- To verify native functions, the bytecode verifier needs the `native_functions` module
- Trying to add `vm_runtime` as a dependency for the bytecode verifier created a cyclic dependency
- `vm_runtime_types` is a lightweight crate that has a small number of dependencies
- `vm_runtime_types` will be used by the bytecode verifier to verify the native function signatures
