+++ Cargo.toml
@@ -46,6 +46,7 @@ members = [
     "language/vm/serializer_tests",
     "language/vm/vm_runtime",
     "language/vm/vm_runtime/vm_cache_map",
+    "language/vm/vm_runtime/vm_runtime_types",
     "language/vm/vm_genesis",
     "libra_node",
     "libra_swarm",
