+++ language/e2e_tests/Cargo.toml
@@ -18,6 +18,7 @@ state_view = { path = "../../storage/state_view" }
 types = { path = "../../types" }
 vm = { path = "../vm" }
 vm_runtime = { path = "../vm/vm_runtime" }
+vm_runtime_types = { path = "../vm/vm_runtime/vm_runtime_types" }
 proptest = "0.9.3"
 proptest-derive = "0.1.1"
 assert_matches = "1.3.0"
