+++ language/vm/vm_genesis/Cargo.toml
@@ -18,6 +18,7 @@ types = { path = "../../../types" }
 vm = { path = "../" }
 vm_cache_map = { path = "../vm_runtime/vm_cache_map" }
 vm_runtime = { path = "../vm_runtime" }
+vm_runtime_types = { path = "../vm_runtime/vm_runtime_types" }
 hex = "0.3.2"
 lazy_static = "1.3.0"
 rand = "0.6.5"
