+++ language/vm/vm_runtime/Cargo.toml
@@ -8,8 +8,6 @@ edition = "2018"
 
 [dependencies]
 bit-vec = "0.6.1"
-bitcoin_hashes = "0.3.0"
-byteorder = "1.3.2"
 hex = "0.3.2"
 lazy_static = "1.3.0"
 proptest = "0.9"
@@ -24,11 +22,11 @@ crypto = { path = "../../../crypto/legacy_crypto"}
 failure = { path = "../../../common/failure_ext", package = "failure_ext" }
 logger = { path = "../../../common/logger" }
 metrics = { path = "../../../common/metrics" }
-nextgen_crypto = { path = "../../../crypto/nextgen_crypto" }
 state_view = { path = "../../../storage/state_view" }
 types = { path = "../../../types" }
 vm = { path = "../" }
 vm_cache_map = { path = "vm_cache_map" }
+vm_runtime_types = { path = "vm_runtime_types" }
 
 [dependencies.prometheus]
 version  = "0.4.2"
