+++ language/vm/vm_runtime/src/frame.rs
@@ -4,7 +4,6 @@
 use crate::{
     bounded_fetch,
     loaded_data::{function::FunctionReference, loaded_module::LoadedModule},
-    value::Local,
 };
 use std::{fmt, marker::PhantomData, mem::replace};
 use vm::{
@@ -12,6 +11,7 @@ use vm::{
     file_format::{Bytecode, CodeOffset, LocalIndex},
     IndexKind,
 };
+use vm_runtime_types::value::Local;
 
 pub struct Frame<'txn, F: 'txn> {
     pc: u16,
