+++ language/vm/vm_runtime/vm_runtime_types/Cargo.toml
@@ -1,0 +1,33 @@
+[package]
+name = "vm_runtime_types"
+version = "0.1.0"
+authors = ["Libra Association <opensource@libra.org>"]
+license = "Apache-2.0"
+publish = false
+edition = "2018"
+
+[dependencies]
+bit-vec = "0.6.1"
+bitcoin_hashes = "0.3.0"
+hex = "0.3.2"
+lazy_static = "1.3.0"
+proptest = "0.9"
+rayon = "1.1"
+rental = "0.5.4"
+tiny-keccak = "1.5.0"
+
+types = { path = "../../../../types" }
+vm = { path = "../../" }
+vm_cache_map = { path = "../vm_cache_map" }
+canonical_serialization = { path = "../../../../common/canonical_serialization" }
+nextgen_crypto = { path = "../../../../crypto/nextgen_crypto" }
+failure = { path = "../../../../common/failure_ext", package = "failure_ext" }
+
+[dependencies.prometheus]
+version  = "0.4.2"
+default-features = false
+features = ["push"]
+
+[features]
+instruction_synthesis = []
+testing = ["types/testing"]
