+++ language/vm/vm_runtime/vm_runtime_types/src/loaded_data/mod.rs
@@ -4,5 +4,5 @@
 //!
 //! This module contains the loaded definition of code data used in runtime.
 
-pub mod function;
-pub mod loaded_module;
+pub mod struct_def;
+pub mod types;
