+++ language/vm/vm_runtime/vm_runtime_types/src/loaded_data/struct_def.rs
