+++ language/vm/vm_runtime/vm_runtime_types/src/native_functions/hash.rs
